changed the generator code to support the optional flags
when element is marked with minOccurs="0" and maxOccurs="1", such element is optional and corresponding property schema schould be marked as optional.
when element is marked with minOccurs="1" and maxOccurs="1", such element is required and corresponding property schema schould be marked as not optional.
when element is marked with maxOccurs greater than 1, such element is considered an element of an array and independent on minOccurs in the array it is considered required.